+++ src/components/About.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/About.css';
+import '../styles/General.css';
 
 const About = () => (
   <div className="main">
