+++ src/components/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../styles/About.css';
+import '../styles/General.css';
 
 const Home = () => (
   <div className="main">
